# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "SDET School MERN store",
    "description": "SWAGGER API Documentation for Students of SDET School"
  },
  "servers": [
    {
      "url": "https://mern-ecommerce.sdet.school/api"
    },
    {
      "url": "http://localhost:4005/api"
    }
  ],
  "paths": {
...
    "/brand/add": {
      "post": {
        "tags": [
          "Brand"
        ],
        "summary": "Add new brand",
        "description": "Add new brand",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBrand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add new brand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/brand/list": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "List all brands",
        "description": "List all brands",
        "responses": {
          "200": {
            "description": "List all brands",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "brands": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AddBrandResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/brand/list/select": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Brands by merchant",
        "description": "Brands by merchant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Brands names and ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/brand": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Brands by merchant",
        "description": "Brands by merchant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Brand by merchant",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "brands": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BrandMerchantResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/brand/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "description": "Brand id",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get Brand",
        "description": "Returns brands products",
        "responses": {
          "200": {
            "description": "Brand's products",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductResponse"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "totalProducts": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brand"
        ],
        "summary": "Update brand",
        "description": "Update brand",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brand updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/brand/delete/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "description": "Brand id"
        }
      ],
      "delete": {
        "tags": [
          "Brand"
        ],
        "summary": "Delete brand by id",
        "description": "Delete brand by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Delete brand by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/merchant/add": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "summary": "Add merchant",
        "description": "Add merchant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMerchant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add Merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddDeleteMerchantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/merchant": {
      "parameters": [
        {
          "in": "query",
          "name": "page",
          "schema": {
            "type": "integer",
            "example": 1
          },
          "required": true
        },
        {
          "in": "query",
          "name": "limit",
          "schema": {
            "type": "integer",
            "example": 10
          },
          "required": true
        }
      ],
      "get": {
        "tags": [
          "Merchant"
        ],
        "summary": "Get all merchants",
        "description": "Get all merchants",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of merchants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/{id}/active": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "description": "Merchant id",
          "required": true
        }
      ],
      "put": {
        "tags": [
          "Merchant"
        ],
        "summary": "Activate/Disable merchant",
        "description": "Activate/Disable merchant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Activate/Disable merchant",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "merchant": {
                    "properties": {
                      "isActive": {
                        "type": "boolean",
                        "required": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update status",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/merchant/delete/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "description": "Merchant id"
        }
      ],
      "delete": {
        "tags": [
          "Merchant"
        ],
        "summary": "Delete merchant",
        "description": "Delete merchant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Delete merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddDeleteMerchantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/item/{slug}": {
      "parameters": [
        {
          "in": "path",
          "name": "slug",
          "required": true,
          "description": "Product slug"
        }
      ],
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get product by slug",
        "description": "Get product by slug",
        "responses": {
          "200": {
            "description": "Get product by slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSlugResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product/list/search/{name}": {
      "parameters": [
        {
          "in": "path",
          "name": "name",
          "required": true,
          "description": "Product name"
        }
      ],
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Search product by name",
        "description": "Search product by name",
        "responses": {
          "200": {
            "description": "Search by product name",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductSearchResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product/list/": {
      "parameters": [
        {
          "in": "query",
          "name": "name",
          "schema": {
            "type": "string",
            "example": "all"
          }
        },
        {
          "in": "query",
          "name": "category",
          "schema": {
            "type": "string",
            "example": "all"
          }
        },
        {
          "in": "query",
          "name": "brand",
          "schema": {
            "type": "string",
            "example": "all"
          }
        },
        {
          "in": "query",
          "name": "min",
          "schema": {
            "type": "integer",
            "example": 1
          }
        },
        {
          "in": "query",
          "name": "max",
          "schema": {
            "type": "integer",
            "example": 2500
          }
        },
        {
          "in": "query",
          "name": "rating",
          "schema": {
            "type": "integer",
            "example": 0
          }
        },
        {
          "in": "query",
          "name": "order",
          "schema": {
            "type": "integer",
            "example": 0
          }
        },
        {
          "in": "query",
          "name": "page",
          "schema": {
            "type": "integer",
            "example": 1
          }
        },
        {
          "in": "query",
          "name": "limit",
          "schema": {
            "type": "integer",
            "example": 10
          }
        },
        {
          "in": "query",
          "name": "sortOrder",
          "schema": {
            "type": "string",
            "example": "{\"_id\":-1}"
          }
        }
      ],
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Search product",
        "description": "Search product",
        "responses": {
          "200": {
            "description": "products",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductResponse"
                      }
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "currentPage": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product/list/brand/{slug}": {
      "parameters": [
        {
          "in": "path",
          "name": "slug",
          "required": true,
          "description": "Brand slug"
        }
      ],
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "List brands by slug",
        "description": "List brands by slug",
        "responses": {
          "200": {
            "description": "List product by brand slug",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductBrandSlugResponse"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "totalProducts": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/list/select": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "List select product",
        "description": "List select product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List select product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListSelectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddBrand": {
        "type": "object",
        "required": [
          "name",
          "description",
          "isActive"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "AddBrandResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "brand": {
            "properties": {
              "isActive": {
                "type": "boolean"
              },
              "merchant": {
                "type": "string"
              },
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "created": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "__v": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}